Initial tuxvdmtool implementation
Ports macvdmtool [1] to Linux using a sysfs interface to the downstream
tipd/cdx321x driver. This is probably not upstreamable but tuxvdmtool is
expected to adapt to interface changes in the AsahiLinux downstream
kernel and eventually the upstream Linux kernel.

UART via /dev/ttySAC0 is unfortunately not stable with `asahi-6.16.5-2`.
The port resets within tens of seconds. Cause for this is probably in
Linux' samsung_tty.c driver.

[1] https://github.com/AsahiLinux/macvdmtool

Signed-off-by: Janne Grunau <j@jannau.net>